cast to int on passed

+++ hscore-minecraft/hscore-minecraft-gui-mask/src/main/java/me/hsgamer/hscore/minecraft/gui/mask/impl/OneTimeAnimatedMask.java
@@ -163,7 +163,7 @@ public class OneTimeAnimatedMask extends BaseMask implements IdentifiedUpdatable
         return;
       }
 
-      long passed = diff / periodMillis;
+      int passed = (int) (diff / periodMillis);
       long remainder = diff % periodMillis;
       lastTickMillis = currentTick - remainder;
